backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
pydantic
certifi
hindsight-client
pytest
httpx
google-generativeai
